[miner] Ministo -> Coordinator
And tentatively run `listener` on a side-thread.
